refactor: PostModal, Feed, OptionsModal
