+++ src/Components/EditModal/EditModal.jsx
@@ -1,9 +1,9 @@
 import { useState, useRef, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { newPost, editPost } from "../Utils";
-import { useClickOustide } from "../../../Hooks/useClickOutside";
+import { newPost, editPost } from "../../features/Post/Utils";
+import { useClickOustide } from "../../Hooks/useClickOutside";
 
-export default function ModalPost({ setShowModal, post }) {
+export default function EditModal({ setShowModal, post }) {
     const [content, setContent] = useState(post?.content || "");
     const contentRef = useRef(null);
     const modalRef = useRef(null);
@@ -34,7 +34,7 @@ export default function ModalPost({ setShowModal, post }) {
     }, [content]);
 
     return (
-        <div className="fixed inset-0 bg-[#00000080] z-10 flex w-full justify-center items-center">
+        <div className="fixed inset-0 bg-modal-background z-10 flex w-full justify-center items-center">
             <div
                 ref={modalRef}
                 className=" pt-4 pb-2 px-3 border-y border-slate-500 rounded-md w-1/3 bg-slate-100 dark:bg-slate-700"
